Don't suppress persisted cache change triggers
Was attempting to only update the stream callbacks when the data actually changed, but this created some issues in my personal scripts, and it's a trivial performance concern.

If this does introduce performance issues, I suggest you rather don't update persisted with no data change, or refactor your calling table.