Ensure that an instantiated table can be rendered multiple times
This fixes a pernicious issue where a table, after having been instantiated and rendered once, would not properly initialize all of its state. The most critical uninitiated element was the connected$ poller. The result of this bug was that any connected table would NOT UPDATE with connected stream updates after the first render.

This issue is fixed by ensuring that on each first render of the table, the connected stream callback is registered. Previously it was only registered upon table instantiation.

Also clarifies some logging & adds an option when fetching persisted data to suppress cache stream change trigger.